question-4/log-analysis-alert-system/requirements.txt
aws-cdk.core
aws-cdk.aws-sqs
aws-cdk.aws-lambda
boto3